LayoutTests/platform/mac/accessibility/aria-divs-not-ignored-expected.txt

text

This tests that a div element with nothing in, but one that has an ARIA role, will not be ignored by accessibility.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS body.childAtIndex(1).role is 'AXRole: AXGroup'
PASS body.childAtIndex(1).subrole is 'AXSubrole: AXApplicationLog'
PASS successfullyParsed is true

TEST COMPLETE
